requirements.txt
Django ~= 2.1
Pillow ~= 6.0
gunicorn ~= 19.9
